saparated FireBase to data and login
