+++ app/src/main/java/com/project/westudentmain/activities/showSettings.java
@@ -4,11 +4,10 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.view.View;
 import android.widget.Button;
 
 import com.example.androidproject.R;
-import com.project.westudentmain.util.FireBase;
+import com.project.westudentmain.util.FireBaseLogin;
 
 public class showSettings extends AppCompatActivity {
     private Button dissconnectbtn,changepassbtn,editprofilebtn,changepicbtn;
@@ -19,7 +18,7 @@ public class showSettings extends AppCompatActivity {
         setContentView(R.layout.activity_show_settings);
         connect_items_by_id();
         dissconnectbtn.setOnClickListener(var -> {
-            FireBase.signOut();
+            FireBaseLogin.signOut();
             finish();
             startActivity(new Intent(showSettings.this, Login.class));
         });
